fix: HTTP Challenge is now also verifying signature
